Fix issue with Flatpickr overriding custom input
Since version 4.6.3 Flatpickr added a fix for an issue
which causes that custom input entered by user is
overwritten with internal datepicker value.

  https://github.com/flatpickr/flatpickr/pull/1879

That means, that if someone changes the date/time
manually and clicks away, the original time would
return.

For now, I've locked the Flatpickr version to 4.6.2
and will issue bug to the developers.