Fix form submit buttons disabling
The actual disable of button caused, that data serialization
that browser does and happens after this callback wouldn't
be able to serialize the submit button data.

That means that backend wouldn't know that a specific submit
button was used and also onClick handlers defined by Nette
form would stop working.

Delaying the execution (even with zero timeout) is sufficient
for browser to do the serialization.

remp/respekt#450